feat: add relations between users and record and adjust related route and database schema

+++ routes/modules/expenses.js
@@ -2,7 +2,7 @@ const express = require('express')
 const router = express.Router()
 const Record = require('../../models/record')
 const Category = require('../../models/category')
-const User = require('../../models/user')
+
 // 取得新增支出表單頁面
 router.get('/new', (req, res) => {
   res.render('new')
@@ -10,21 +10,20 @@ router.get('/new', (req, res) => {
 // 送出新增支出表單內容
 router.post('/', async (req, res) => {
   const { name, date, category, amount } = req.body
-  Promise.all([
-    Category.findOne({name: category}).lean(),
-    User.create({name: 'test', email: 'test@example.com', password: '12345'}),
-  ])
-  .then(([category, user]) => {
-    const userId = user._id
-    const categoryId = category._id
-    return Record.create({
-      name,
-      date,
-      amount,
-      categoryId,
-      userId
+  const userId = req.user._id
+  
+  Category.findOne({name: category})
+    .lean()
+    .then(category => {
+      const categoryId = category._id
+      return Record.create({
+        name,
+        date,
+        amount,
+        categoryId,
+        userId
+      })
     })
-  })
   .then(() => {
     res.redirect('/')
   })
@@ -33,16 +32,12 @@ router.post('/', async (req, res) => {
   // try {
   //   const { name, date, category, amount } = req.body
   //   const categoryDoc = await Category.findOne({ name: category })
-  //   const userId = await User.create({
-  //     name: 'test',
-  //     email: 'test@example.com',
-  //     password: '12345'
-  //   })
+  //   const userId = req.user._id
   //   Record.create({
   //     name,
   //     date,
   //     amount,
-  //     userId: userId._id,
+  //     userId,
   //     categoryId: categoryDoc._id
   //   })
   //   res.redirect('/')
@@ -53,9 +48,10 @@ router.post('/', async (req, res) => {
 // 取得編輯支出費用頁面
 router.get('/:id/edit', async(req, res) => {
   const _id = req.params.id
+  const userId = req.user._id
   Promise.all([
     Category.find().lean(),
-    Record.findOne({ _id }).populate('categoryId').lean()
+    Record.findOne({ _id, userId }).populate('categoryId').lean()
   ])
   .then(([categories, record]) => {
     const categoryOptions = categories.map(category => ({
@@ -69,8 +65,9 @@ router.get('/:id/edit', async(req, res) => {
   // await/async的寫法
   // try{
   //   const _id = req.params.id
+  //   const userId = req.user._id
   //   const categories = await Category.find().lean()
-  //   const record = await Record.findOne({_id }).populate('categoryId').lean()
+  //   const record = await Record.findOne({ _id, userId }).populate('categoryId').lean()
   //   const categoryOptions = categories.map(category => ({
   //     name: category.name,
   //     isSelected: category.name === record.categoryId.name
@@ -85,9 +82,10 @@ router.get('/:id/edit', async(req, res) => {
 router.put('/:id', async (req, res) => {
   const { name, date, category, amount } = req.body
   const _id = req.params.id
+  const userId = req.user._id
   Promise.all([
     Category.findOne({name: category}).lean(),
-    Record.findOne({_id})         // 這邊不能加 lean()
+    Record.findOne({ _id, userId })         // 這邊不能加 lean()
   ])
   .then(([category, record]) => {
     const categoryId = category._id
@@ -105,8 +103,9 @@ router.put('/:id', async (req, res) => {
   // try {
   //   const { name, date, category, amount } = req.body
   //   const _id = req.params.id
+  //   const userId = req.user._id
   //   const categoryDoc = await Category.findOne({name: category})
-  //   const record = await Record.findById( {_id} )
+  //   const record = await Record.findOne({ _id, userId })
   //   record.name = name,
   //   record.date = date,
   //   record.categoryId = categoryDoc._id,
@@ -120,13 +119,15 @@ router.put('/:id', async (req, res) => {
 // 刪除特定支出費用
 router.delete('/:id', async (req, res) => {
   const _id = req.params.id
-  Record.findByIdAndDelete({_id})
+  const userId = req.user._id
+  Record.findByIdAndDelete({ _id, userId })
     .then(() => res.redirect('/'))
     .catch(error => console.log(error))
   // async/ await  
   // try {
   //   const _id = req.params.id
-  //   await Record.findByIdAndDelete({_id})
+  //   const userId = req.user._id
+  //   await Record.findByIdAndDelete({ _id, userId })
   //   res.redirect('/')
   // } catch (error) {
   //   console.log(error)
